revert resource relocation
