+++ go.mod
@@ -3,18 +3,24 @@ module github.com/giantswarm/kvm-operator
 go 1.14
 
 require (
+	github.com/giantswarm/apiextensions v0.4.20
 	github.com/giantswarm/apiextensions/v2 v2.6.2
 	github.com/giantswarm/certs/v3 v3.1.0
 	github.com/giantswarm/errors v0.2.3
+	github.com/giantswarm/k8sclient/v2 v2.1.0
 	github.com/giantswarm/k8sclient/v4 v4.0.0
 	github.com/giantswarm/k8scloudconfig/v8 v8.0.4
 	github.com/giantswarm/microendpoint v0.2.0
 	github.com/giantswarm/microerror v0.2.1
 	github.com/giantswarm/microkit v0.2.2
 	github.com/giantswarm/micrologger v0.3.3
+	github.com/giantswarm/operatorkit v1.2.0 // indirect
 	github.com/giantswarm/operatorkit/v2 v2.0.1
+	github.com/giantswarm/randomkeys v0.2.0
 	github.com/giantswarm/randomkeys/v2 v2.0.0
+	github.com/giantswarm/statusresource v0.4.0
 	github.com/giantswarm/statusresource/v2 v2.0.0
+	github.com/giantswarm/tenantcluster v0.2.0
 	github.com/giantswarm/tenantcluster/v3 v3.0.0
 	github.com/giantswarm/versionbundle v0.2.0
 	github.com/google/go-cmp v0.5.2
